Put the unlocked mods selector back in the mod picker. Remove the max width of a generated set. Calculate the plus5mods in the loadout builder class itself.
