+++ src/app/loadout-builder/generated-sets/GeneratedSet.m.scss
@@ -32,7 +32,7 @@
   flex-direction: row;
   flex-wrap: wrap;
   padding: 0;
-  max-width: 800px;
+  //max-width: 800px;
 
   @include phone-portrait {
     justify-content: space-around;
